% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benth.met.R
\name{benth.met}
\alias{benth.met}
\title{Benthic Metric Calculation}
\usage{
benth.met(x, tax.fields = 2, site.fields, HBI = NULL)
}
\arguments{
\item{x}{data.frame with sampling events in rows and taxa in columns. No row or column names should be defined}

\item{tax.fields}{The number of rows in x used for taxa names (will be concatenated in output)}

\item{site.fields}{The number of columns in x used for site names (will be concatenated in output)}

\item{HBI}{Custom sensitivity values for HBI calculation. Must follow format of data(HBI1,envir = environment())}
}
\value{
$Summary.Metrics - Calculated indicator metrics

$Raw.Data - Raw taxon data

$Taxa.List - Concatenated taxon names

$Site.List - Concatenated site names
}
\description{
Calculation of a variety of metrics for determing impairment of Benthic Macroinvertebrate Communities.
}
\examples{
data(YKBioData,envir = environment())
benth.met(YKBioData,2,2)
}
\keyword{Benthic}
\keyword{Metrics}

