% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting Functions.R
\name{boxplot.tsa.object}
\alias{boxplot.tsa.object}
\title{z-score Boxplots}
\usage{
\method{boxplot}{tsa.object}(tsa.object, ...)
}
\arguments{
\item{tsa.object}{An object from tsa.test()}

\item{...}{Arguments passed to boxplot()}
}
\description{
Boxplots of z-scores used in Test Site Analysis. Metrics used in calculation of TSA are circled, metrics that
contribute significantly to an impairment score are marked with an * along the x axis.
}
\examples{
#load datasets
data(YKEnvData,envir = environment()) #Biological dataset
data(YKBioData,envir = environment()) #Environmental dataset

#Calculate indicator metrics from raw biological data
bio.data.test<-benth.met(YKBioData,2,2)

#Extract just the summary metrics
bio.data<-bio.data.test$Summary.Metrics

#standardize row names between datasets
rownames(YKEnvData)<-bio.data.test$Site.List

#Match a test site (#201) to the nearest neighbour reference set
nn.sites<-site.match(YKEnvData[201,-c(1)],YKEnvData[1:118,-c(1)],k=NULL,adaptive=T)

#Calculate additional metrics based on selected Reference sites
taxa.data<-add.met(Test=bio.data.test$Raw.Data[201,],Reference=bio.data.test$Raw.Data[names(nn.sites$final.dist),])

#TSA test of indicator metrics at test site and reference sites selected used site.match()
tsa.results<-tsa.test(Test=taxa.data[nrow(taxa.data),],Reference=taxa.data[names(nn.sites$final.dist),],distance=nn.sites$final.dist, outlier.rem=T, m.select=T)
tsa.results

Evaluate Results
boxplot(tsa.results)
}
\keyword{Analysis,}
\keyword{Boxplot}
\keyword{Site}
\keyword{Test}

