% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CEFI}
\alias{CEFI}
\title{Canadian Ecological Flow Index}
\format{
An object of class \code{data.frame} with 55 rows and 3 columns.
}
\usage{
CEFI
}
\description{
Canadian Ecological Flow Index
}
\keyword{datasets}
