% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HBI_RAW}
\alias{HBI_RAW}
\title{Hilsenhoff Biotic Index}
\format{
An object of class \code{data.frame} with 1283 rows and 3 columns.
}
\usage{
HBI_RAW
}
\description{
Hilsenhoff Biotic Index
}
\keyword{datasets}
