% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.select.UI.R
\name{metric.select.UI}
\alias{metric.select.UI}
\title{Indicator metric selection}
\usage{
metric.select.UI(Test, Reference, outlier.rem = T, rank = F,
  outbound = 0.1)
}
\arguments{
\item{Test}{Vector containing metric scores at the test site. Should be a single row from \code{benth.met} or \code{add.met}.}

\item{Reference}{Data frame of metric scores at the reference sites. Should be output from \code{benth.met} or \code{add.met}.}

\item{outbound}{Used if outlier.rem=T A numeric value between 0 and 1 indicating the outlier boundary for defining values as final outliers (default to 0.1)}

\item{Rank}{Use rank differences in metric selection}
}
\value{
$Best.Metrics - Vector containing the final selected indicator metrics

$Indicative.Metrics - Vector containing all metrics that indicate impairment

$raw.data - Data frame containing only selected best metrics

$ref.sites - Vector containing input reference site names

$outlier.ref.sites - Vector containing sites removed as potential outliers
}
\description{
Determines which indicator metrics which best differentiate the test site from its nearest-neighbbour reference sites. Metrics that indicate impairment will be
used preferentially.
}
\details{
A interative selection algorithm is used as follows:

1. The first metric selected for the final set is the one which displayes the greatest distance from the Reference condition mean

2. Metrics with a pearson correlation greater than 0.7 to (any of) the selected metric(s) are excluded from further steps

3. The ranked departure of remaining metrics is divided by the (maximum) correlation with the metric(s) previously included in the analysis

4. The metric with the greatest score is selected for inclusion in the final set

5. Return to step 2 until the number of selected metrics is equal to the greater of 4 or 1/5 the number of Reference sites

If no metrics or too few metrics demonstrate impairment, the following metrics are included until the maximum is reached:
Richness, Percent Dominance, HBI, Percent EPT.
}
\examples{
data(YKBioData,envir = environment())
bio.data<-benth.met(YKBioData,2,2)$Summary.Metrics
nn.refsites<- c("075-T-1", "019-T-1","003-T-1","076-T-1","071-T-1","022-T-1","074-T-1",
"002-T-1","004-T-1","073-T-1","186-T-1","062-T-1","005-T-1","025-T-1",
"187-T-1","023-T-1","193-T-1","192-T-1","196-T-1","194-T-1")
metric.select(bio.data[201,],bio.data[nn.refsites,])
}
\keyword{Benthic}
\keyword{Metrics}
