% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting Functions.R
\name{plot.match.object}
\alias{plot.match.object}
\title{Plot of nearest-neighbour reference site distances to test site}
\usage{
plot.match.object(match.object)
}
\arguments{
\item{site.match.object}{An object from \code{\link[BenthicAnalysis]{"site.match"}}}
}
\description{
Plot of nearest-neighbour reference site distances to test site
}
\examples{
#load datasets
data(YKEnvData,envir = environment()) #Biological dataset
data(YKBioData,envir = environment()) #Environmental dataset

#Calculate indicator metrics from raw biological data
bio.data.test<-benth.met(YKBioData,2,2)

#Extract just the summary metrics
bio.data<-bio.data.test$Summary.Metrics

#standardize row names between datasets
rownames(YKEnvData)<-bio.data.test$Site.List

#Match a test site (#201) to the nearest neighbour reference set
nn.sites<-site.match(YKEnvData[201,-c(1)],YKEnvData[1:118,-c(1)],k=NULL,adaptive=T)

Evaluate Results
plot(nn.sites)
}
\keyword{ANNA,}
\keyword{RDA-ANNA}

