% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting Functions.R
\name{sitematch.plot}
\alias{sitematch.plot}
\title{Scatterplot of ANNA and RDA-ANNA site matching functions}
\usage{
sitematch.plot(match.object, axis = c(1, 2))
}
\arguments{
\item{axis}{A two character vector indicating which axis to plot}

\item{site.match.object}{An object from \code{\link[BenthicAnalysis]{"site.match"}}}
}
\description{
Scatterplot of ANNA and RDA-ANNA site matching functions.
}
\examples{
#load datasets
data(YKEnvData,envir = environment()) #Biological dataset
data(YKBioData,envir = environment()) #Environmental dataset

#Calculate indicator metrics from raw biological data
bio.data.test<-benth.met(YKBioData,2,2)

#Extract just the summary metrics
bio.data<-bio.data.test$Summary.Metrics

#standardize row names between datasets
rownames(YKEnvData)<-bio.data.test$Site.List

#Match a test site (#201) to the nearest neighbour reference set
nn.sites<-site.match(YKEnvData[201,-c(1)],YKEnvData[1:118,-c(1)],k=NULL,adaptive=T)

Evaluate Results
sitematch.plot(nn.sites)
}
\keyword{ANNA,}
\keyword{RDA-ANNA}

