% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trait.feeding}
\alias{trait.feeding}
\title{Lookup Table of feeding traits}
\format{
An object of class \code{data.frame} with 1924 rows and 2 columns.
}
\usage{
trait.feeding
}
\description{
Lookup Table of feeding traits
}
\keyword{datasets}
