% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trait.habit}
\alias{trait.habit}
\title{Lookup Table of habitat traits}
\format{
An object of class \code{data.frame} with 1583 rows and 2 columns.
}
\usage{
trait.habit
}
\description{
Lookup Table of habitat traits
}
\keyword{datasets}
