% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{tsa.conf}
\alias{tsa.conf}
\title{Test Site Analysis confidence}
\usage{
tsa.conf(match.object, tsa.object, Reference = NULL, benth.metric = NULL,
  levels = c(1, 5, 10, 20, 30, 40), reps = 100)
}
\arguments{
\item{match.object}{An object from \code{\link[BenthicAnalysis]{site.match}}}

\item{tsa.object}{An object from \code{\link[BenthicAnalysis]{tsa.test}}}

\item{Reference}{Metric scores of Reference Set}

\item{benth.metric}{An object from \code{\link[BenthicAnalysis]{benth.met}}}

\item{levels}{Percent of Reference sites to remove}

\item{reps}{Replicate jacknife samples to assess}
}
\value{
$summary.results
}
\description{
Assess the confidence of TSA results based on jacknifing entire reference set.
}
\examples{
tsa.conf()
}
\keyword{Analysis,}
\keyword{Jacknife}
\keyword{Site}
\keyword{Test}

