% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{tsa.zscore}
\alias{tsa.zscore}
\title{Z-Score Calculation}
\usage{
tsa.zscore(Test, Reference, na.cutoff = 0.7)
}
\arguments{
\item{Test}{Data frame of metric scores at the test site.}

\item{Reference}{Data frame of metric scores at the reference sites}

\item{na.cutoff}{A value between 0-1 indicating the percent of the reference set that can contain NAs for any metric. NAs are replaced by the mean value.}
}
\description{
This calculates z-scores centered on Reference Sites. A certain number of NAs can be permitted in the reference set for each metric but
metrics with NAs at the test site are automatically removed
}
\examples{
tsa.zscore()
}
\keyword{z-score}

